Feat : 앱 종료하는 api 연결
